[ゼロから作るDeep Learning]勉強中

+++ dlfs/ch06/hyperparameter_optimization.py
@@ -1,5 +1,6 @@
 # coding: utf-8
 import sys, os
+
 sys.path.append(os.pardir)  # 親ディレクトリのファイルをインポートするための設定
 import numpy as np
 import matplotlib.pyplot as plt
@@ -16,7 +17,7 @@ t_train = t_train[:500]
 
 # 検証データの分離
 validation_rate = 0.20
-validation_num = x_train.shape[0] * validation_rate
+validation_num = int(x_train.shape[0] * validation_rate)
 x_train, t_train = shuffle_dataset(x_train, t_train)
 x_val = x_train[:validation_num]
 t_val = t_train[:validation_num]
@@ -58,11 +59,11 @@ col_num = 5
 row_num = int(np.ceil(graph_draw_num / col_num))
 i = 0
 
-for key, val_acc_list in sorted(results_val.items(), key=lambda x:x[1][-1], reverse=True):
-    print("Best-" + str(i+1) + "(val acc:" + str(val_acc_list[-1]) + ") | " + key)
+for key, val_acc_list in sorted(results_val.items(), key=lambda x: x[1][-1], reverse=True):
+    print("Best-" + str(i + 1) + "(val acc:" + str(val_acc_list[-1]) + ") | " + key)
 
-    plt.subplot(row_num, col_num, i+1)
-    plt.title("Best-" + str(i+1))
+    plt.subplot(row_num, col_num, i + 1)
+    plt.title("Best-" + str(i + 1))
     plt.ylim(0.0, 1.0)
     if i % 5: plt.yticks([])
     plt.xticks([])
